+++ examples/angular/src/app/header/header.ts
@@ -1,31 +1,35 @@
-import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output, ViewEncapsulation } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input, output, signal } from '@angular/core';
 import { Language, languages } from '@six-group/ui-library';
+import { UiLibraryAngularModule } from '@six-group/ui-library-angular';
 
 @Component({
   selector: 'app-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.scss'],
+  imports: [UiLibraryAngularModule],
+  templateUrl: './header.html',
+  styleUrl: './header.css',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  encapsulation: ViewEncapsulation.ShadowDom,
 })
-export class HeaderComponent {
-  @Input() taskCount: number = 0;
-  @Input() open = false;
+export class Header {
+  // inputs
+  taskCount = input.required<number>();
+  open = input(false);
 
-  @Output() leftSidebarToggled = new EventEmitter();
-  @Output() rightSidebarToggled = new EventEmitter();
+  // outputs
+  leftSidebarToggled = output();
+  rightSidebarToggled = output();
 
-  language: Language;
+  // local state
+  language = signal<Language>('de');
   languages: Language[] = [...languages];
   apps = ['Application 1', 'Application 2', 'Application 3', 'Application 4'];
-  currentApp = 'Application 1';
+  currentApp = signal('Application 1');
 
   constructor() {
     let lang = localStorage.getItem('six-lang') as Language;
     if (lang == null || !this.languages.includes(lang)) {
       lang = 'de';
     }
-    this.language = lang;
+    this.language.set(lang);
     document.documentElement.lang = lang;
   }