+++ examples/angular/src/app/home/home.ts
@@ -1,0 +1,11 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { UiLibraryAngularModule } from '@six-group/ui-library-angular';
+
+@Component({
+  selector: 'app-home',
+  imports: [UiLibraryAngularModule],
+  templateUrl: './home.html',
+  styleUrl: './home.css',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class Home {}