+++ libraries/ui-library/src/components/six-tab/six-tab.tsx
@@ -88,6 +88,7 @@ export class SixTab {
           aria-disabled={this.disabled ? 'true' : 'false'}
           aria-selected={this.active ? 'true' : 'false'}
           tabindex={this.disabled || !this.active ? '-1' : '0'}
+          title={this.host.innerHTML}
         >
           <slot />
           {this.closable && (