+++ docs/changelog.md
@@ -54,6 +54,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
   value of the `multiple` property
 - Upgraded Stencil to latest release and upgraded Vue output target
 - Upgraded React output target to version 1.0.0
+- Upgraded Angular output target to version 0.10.2
 
 ## 4.3.2 - 2025-02-13