+++ docs/changelog.md
@@ -25,7 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - `six-select` : support the use of prefix/suffix slot of the `six-menu-item` to display icons.
 - Added support for aria-attributes in JSX typings through upgrade of an underlying library
 - When enabled, added automatic `required` flag on input components using `FormControl` and
-  `Validators.required`. See
+  `Validators.required` on Angular. See
   [docs](./guide/angular.md#automatic-required-when-using-formcontrol-and-validatorsrequired) for
   more details.