+++ docs/changelog.md
@@ -17,7 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - [Documentation](guide/angular.md) for using web components with Angular's standalone bootstrapping
 - `ActiveSidebarItemGroupDirective` and `ActiveSidebarItemDirective`: Added optional angular sidebar
   helper directives. See the [Angular guide](guide/angular.md) for more information.
-- Added nested sidbar menu to the angular demo application to showcase the helper directives
+- Added nested sidebar menu to the angular demo application to showcase the helper directives
 - `six-header-dropdown`: Added `filter` and `filterPlaceholder` properties.
 - `six-header-menu-button`: Added `caret`, `disabled`, `loading`, `submit` and `reset` properties.
 - `six-header-menu-button`: Added `suffix` and `prefix` slots.
@@ -37,6 +37,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - `six-timepicker`: removed unnecessary bottom padding
 - `six-button` : simplify use of prefix and suffix icons
 - `six-sidebar-item-group`: fixed spacing for subitems
+- Vue `v-model` issue through an update of the underlying library
 
 ### Removed