+++ docs/changelog.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
   more details.
 - `six-file-upload`: added `code` attribute to `SixFileUploadFailurePayload`, which specifies the
   type of error that has occurred
+- New six-breadcrumbs and six-breadscrumb-item component
 
 ### Fixed